removed the ab3d.dxengine.oculuswrap...
i wanted this to have the 60 days trial using the ab3d.dxengine.OculusWrap for people to use the ab3d.dxengine.oculusWrap. i think i should have asked the permission before doing this. but i was kinda pushing towards promoting the use of the ab3d.dxengine.OculusWrap instead of not using it. i thought that the dll ab3d.dxengine.OculusWrap dll could just be activated at the same time as the ab3d.dxengine nugget when that one activates...

You will have to go clone the repository here https://github.com/ab4d/Ab3d.OculusWrap and build the dlls separately for yourselfves. If the ab3d.dxengine.oculusWrap would maybe be provided in the future with a nugget you won't have to do those steps.

1. Clone the github repository here: https://github.com/ab4d/Ab3d.OculusWrap
2. Build the ab3d.OculusWrap solution first with the FrameWork 4.5. it works in the framework 4.7.2 also.
3. Then build the solution ab3d.DXEngine.OculusWrap.
4. take both the ab3d.OculusWrap.dll and the ab3d.DXEngine.OculusWrap and insert that in the reference of sc_core_systems and SCCoreSystems and the solution sccsv10 and this one sccsv11 and that one too sccsVD4VE. Those DLLs will make the projects work. after inserting those references, rebuild your projects and this should take care of restoring the nugget packages for the other dlls to load.

i will build a solution without those two dlls soon as i've already done it in the past, but my goal was to promote with and without the dxxengine to promote the ab3d.dxengine, as it's been a while already that i learned to not use it. My console directx version, isn't with wpf, but i somehow cannot activate the voice recognition of windows 10 on it and i use the dxengine all the time only for the virtual reality dx engine.